requirements.txt
FastAPI
uvicorn
pydantic
azure-data-tables
azure-identity
pandas